+++ yss/instance/gd32/instance_dac_gd32f1.cpp
@@ -32,12 +32,12 @@ static void setDac1IntEn(bool en)
 	nvic.setDac1En(en);
 }
 
-static unsigned long getDac1ClockFreq(void)
+static uint32_t getDac1ClockFreq(void)
 {
 	return clock.getApb1ClkFreq();
 }
 
-Dac dac1((volatile unsigned int*)DAC, setDac1ClockEn, setDac1IntEn, getDac1ClockFreq);
+Dac dac1((volatile uint32_t*)DAC, setDac1ClockEn, setDac1IntEn, getDac1ClockFreq);
 
 #endif
